requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.3
supabase
python-dotenv
undetected-chromedriver
psycopg2